[GH-0] Reimplement MainWindow without using singleton pattern.
Instead of keeping and maintaining the singleton instance inside the MainWindow class, store it in the window's extra data, so it can be later retrieved in the window procedure dispatcher, which then routes the messages to the real window procedure which is a normal method. This way the window procedure can access the MainWindow instance casually.

Another advantage is that the MainWindow's destructor is invoked automatically when the program closes.