[GH-14] Display all combinations
This PR adds a nice feature to Hotkey Detective which is showing all pressed key combinations, even if it's not taken by any process. The free key combinations display "Unassigned" in the "Process path" column. This is mostly for feedback purposes, so users can see that anything is happening at all, but I also improved the mechanics of identifying key combinations, and made it more C++ friendly.

* [GH-0] Add .clang-format and .clang-tidy.

* [GH-14] Add "setMainWindowKeyboardHook" declaration, and reformat file.

* [GH-14] Reformat MainWindow.cpp.

* [GH-0] Add debugging utils.

* [GH-0] Allocate debug console in window main function.

Remove useless main.h file, and reformat main.cpp.

* [GH-14] Implement classes needed to resolve and keep keystrokes.

This changeset comprises data types and implementations which will be used to translate keystrokes from machine language to human language, and will help to identify and track them.

* [GH-0] Reimplement MainWindow without using singleton pattern.

Instead of keeping and maintaining the singleton instance inside the MainWindow class, store it in the window's extra data, so it can be later retrieved in the window procedure dispatcher, which then routes the messages to the real window procedure which is a normal method. This way the window procedure can access the MainWindow instance casually.

Another advantage is that the MainWindow's destructor is invoked automatically when the program closes.

* [GH-14] Refactor KeySequence to reflect the recent changes.

Keystrokes are no longer obtained from the hook, instead the WM_KEYDOWN and WM_KEYUP are used.

* [GH-14] Implement "clear" method in KeySequence.

When translating the keystroke into a human-friendly string, do not distinguish between left and right modifiers.

* [GH-14] Process key related messages and add them to the sequence.

Detected sequence are now printed to the debug console, but later will be displayed in the table.

When WM_KILLFOCUS message arrives, the sequencer is cleared because that means the window has lost the keyboard focus, so any ongoing key presses should be discarded.

* [GH-14] Implement creation of key sequences from WM_HOTKEY messages.

* [GH-14] Adapt HotkeyTable.addEntry to take parameters of type wstring.

Adapt rest of the code to use the new way of storing strings.

Refactor the code to conform to the new code style.

* [GH-14] Use KeySequence to decode and show hotkeys in the table.

Also add a row to the table for every unassigned key combination.

* [GH-14] Make HotkeyDetective topmost window.

* [GH-14] Remove now obsolete keystrokeToString function.