[GH-5] Move key interception to the main message loop.
This fixes a hidden bug whereby keys are not detected when any of the table entries is selected. This is because the table receives the WM_KEYUP(DOWN) messages instead of the main window because it has a focus. The solution was to move the key interception into the main message loop, and then decide whether to dispatch it or not.