[GH-5] Unload DLL on application exit
This PR fixes the problem described in GH-5 which is inability to remove hotkey_hook.dll after opening and closing Hotkey Detective.

* [GH-5] Rewrite the DLL code to C++ syntax.

* [GH-5] Enable unicode support for DLL and link Shlwapi library.

Since the main application works in unicode mode, it's good to be consistent, so the library should work in unicode mode too.

The additional Shlwapi library is for string operations.

* [GH-5] Unload DLL by using a dedicated thread and a global event.

The main application and the hook DLL are synchronized by using a global event object. As soon as the main application is about to be closed, the event is signalized, what unblocks a tiny thread created within DLL. The only purpose of this thread is to invoke FreeLibraryAndExitThread.

* [GH-5] Implement a method for removing hooks.

Hooks must be removed prematurely, before the terminating event is signalized.

Performed small refactoring, and replaced the type of the shared data to HkdHookData struct.

* [GH-5] Wait for the DLL to unload before closing the window.

This is just a prototype that must be replaced with a real feedback to the user.

* [GH-5] Add dummy VS solution for editing resources.

Visual Studio's built-in resource editor doesn't work very well with .rc files which are not assigned to any project. Since there are no better tools to edit .rc files, the dummy project has been added for the pure sake of forcing VS to cooperate.

* [GH-5] Implement a procedure of the pre-shutdown dialog.

The procedure takes care of polling the status of freeing the hook DLL, and updating the progress bar.

* [GH-5] Move key interception to the main message loop.

This fixes a hidden bug whereby keys are not detected when any of the table entries is selected. This is because the table receives the WM_KEYUP(DOWN) messages instead of the main window because it has a focus. The solution was to move the key interception into the main message loop, and then decide whether to dispatch it or not.

* [GH-5] Show the pre-shutdown dialog when freeing the hook DLL.

* [GH-0] Ignore "project.aps", it is generated during build.

Also ignore .vs directory.