[GH-0] Register custom window message to communicate with DLL.
Using WM_NULL to exchange messages between hooked processes and the main application window was a dirty hack. There was a risk of receiving WM_NULL message which didn't come from the hooked process. The better approach is to register a custom, system-wide message which can be safely uses to exchange message between the main application window and all hooked processes.